+++ LICENSE
@@ -1,6 +1,6 @@
 The MIT License
 
-Copyright (c) 2019 Samuel Wilder, https://github.com/Pebaz
+Copyright (c) 2022 Samuel Wilder, https://github.com/Pebaz
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of
 this software and associated documentation files (the "Software"), to deal in
