--- gen_docs.sh
@@ -1,2 +1,0 @@
-pdoc --html --html-dir docs nimporter.py --overwrite
-mv docs/nimporter.m.html docs/index.html
